requirements.txt
decorator==4.0.10
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
konlpy==0.4.4
numpy==1.11.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
protobuf==3.1.0
ptyprocess==0.5.1
Pygments==2.1.3
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
tensorflow-gpu==0.12.0rc0
traitlets==4.3.1
wcwidth==0.1.7
